pyproject.toml
[project]
name = "teller-integration"
version = "0.1.0"
description = ""
authors = [
    {name = "coulter-fullgood",email = "coulter@fullgoodfulfillment.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "yfinance (>=0.2.65,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.4.1"
mypy = "^1.17.1"

